refactor: Extract FallbackFavicon into a separate component
